pyproject.toml
[project]
name = "subtitler-container"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "openai-whisper>=20250625",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
]
